+++ internal/gatewayapi/securitypolicy.go
@@ -719,89 +719,6 @@ func (t *Translator) buildExtAuth(
 }
 
 // TODO: zhaohuabing combine this function with the one in the route translator
-// validateExtServiceBackendReference validates the backend reference for an
-// external service referenced by an EG policy.
-// This can also be used for the other external services deployed in the cluster,
-// such as the external processing filter, gRPC Access Log Service, etc.
-// It checks:
-//  1. The group is nil or empty, indicating the core API group.
-//  2. The kind is Service.
-//  3. The port is specified.
-//  4. The service exists and the specified port is found.
-//  5. The cross-namespace reference is permitted by the ReferenceGrants if the
-//     namespace is different from the policy's namespace.
-func (t *Translator) validateExtServiceBackendReference(
-	backendRef *gwapiv1.BackendObjectReference,
-	ownerNamespace string,
-	resources *Resources) error {
-
-	// These are sanity checks, they should never happen because the API server
-	// should have caught them
-	if backendRef.Group != nil && *backendRef.Group != "" {
-		return errors.New(
-			"group is invalid, only the core API group (specified by omitting" +
-				" the group field or setting it to an empty string) is supported")
-	}
-	if backendRef.Kind != nil && *backendRef.Kind != KindService {
-		return errors.New("kind is invalid, only Service (specified by omitting " +
-			"the kind field or setting it to 'Service') is supported")
-	}
-	if backendRef.Port == nil {
-		return errors.New("a valid port number corresponding to a port on the Service must be specified")
-	}
-
-	// check if the service is valid
-	serviceNamespace := NamespaceDerefOr(backendRef.Namespace, ownerNamespace)
-	service := resources.GetService(serviceNamespace, string(backendRef.Name))
-	if service == nil {
-		return fmt.Errorf("service %s/%s not found", serviceNamespace, backendRef.Name)
-	}
-	var portFound bool
-	for _, port := range service.Spec.Ports {
-		portProtocol := port.Protocol
-		if port.Protocol == "" { // Default protocol is TCP
-			portProtocol = v1.ProtocolTCP
-		}
-		// currently only HTTP and GRPC are supported, both of which are TCP
-		if port.Port == int32(*backendRef.Port) && portProtocol == v1.ProtocolTCP {
-			portFound = true
-			break
-		}
-	}
-
-	if !portFound {
-		return fmt.Errorf(
-			"TCP Port %d not found on service %s/%s",
-			*backendRef.Port, serviceNamespace, string(backendRef.Name),
-		)
-	}
-
-	// check if the cross-namespace reference is permitted
-	if backendRef.Namespace != nil && string(*backendRef.Namespace) != "" &&
-		string(*backendRef.Namespace) != ownerNamespace {
-		if !t.validateCrossNamespaceRef(
-			crossNamespaceFrom{
-				group:     egv1a1.GroupName,
-				kind:      KindSecurityPolicy,
-				namespace: ownerNamespace,
-			},
-			crossNamespaceTo{
-				group:     GroupDerefOr(backendRef.Group, ""),
-				kind:      KindDerefOr(backendRef.Kind, KindService),
-				namespace: string(*backendRef.Namespace),
-				name:      string(backendRef.Name),
-			},
-			resources.ReferenceGrants,
-		) {
-			return fmt.Errorf(
-				"backend ref to %s %s/%s not permitted by any ReferenceGrant",
-				KindService, *backendRef.Namespace, backendRef.Name)
-		}
-	}
-	return nil
-}
-
-// TODO: zhaohuabing combine this function with the one in the route translator
 func (t *Translator) processExtServiceDestination(
 	backendRef *gwapiv1.BackendObjectReference,
 	ownerNamespace string,
