+++ internal/xds/translator/oidc.go
@@ -62,13 +62,6 @@ func (*oidc) patchHCM(mgr *hcmv3.HttpConnectionManager, irListener *ir.HTTPListe
 			continue
 		}
 
-		// skip if the filter already exists
-		for _, existingFilter := range mgr.HttpFilters {
-			if filter.Name == existingFilter.Name {
-				continue
-			}
-		}
-
 		mgr.HttpFilters = append(mgr.HttpFilters, filter)
 	}
